add new options to pub command
